Rename the `isAfterCommentLine` util to `isAfterComment`
