+++ lib/rules/at-rule-empty-line-before/index.js
@@ -4,7 +4,7 @@ const _ = require("lodash");
 const addEmptyLineBefore = require("../../utils/addEmptyLineBefore");
 const getPreviousNonSharedLineCommentNode = require("../../utils/getPreviousNonSharedLineCommentNode");
 const hasEmptyLine = require("../../utils/hasEmptyLine");
-const isAfterCommentLine = require("../../utils/isAfterCommentLine");
+const isAfterComment = require("../../utils/isAfterComment");
 const isBlocklessAtRuleAfterBlocklessAtRule = require("../../utils/isBlocklessAtRuleAfterBlocklessAtRule");
 const isBlocklessAtRuleAfterSameNameBlocklessAtRule = require("../../utils/isBlocklessAtRuleAfterSameNameBlocklessAtRule");
 const isFirstNested = require("../../utils/isFirstNested");
@@ -97,7 +97,7 @@ const rule = function(expectation, options, context) {
       // Optionally ignore the expectation if a comment precedes this node
       if (
         optionsMatches(options, "ignore", "after-comment") &&
-        isAfterCommentLine(atRule)
+        isAfterComment(atRule)
       ) {
         return;
       }
