+++ lib/rules/comment-empty-line-before/index.js
@@ -2,7 +2,7 @@
 
 const addEmptyLineBefore = require("../../utils/addEmptyLineBefore");
 const hasEmptyLine = require("../../utils/hasEmptyLine");
-const isAfterCommentLine = require("../../utils/isAfterCommentLine");
+const isAfterComment = require("../../utils/isAfterComment");
 const isFirstNested = require("../../utils/isFirstNested");
 const isFirstNodeOfRoot = require("../../utils/isFirstNodeOfRoot");
 const isSharedLineComment = require("../../utils/isSharedLineComment");
@@ -60,7 +60,7 @@ const rule = function(expectation, options, context) {
       // Optionally ignore newlines between comments
       if (
         optionsMatches(options, "ignore", "after-comment") &&
-        isAfterCommentLine(comment)
+        isAfterComment(comment)
       ) {
         return;
       }
