+++ lib/rules/custom-property-empty-line-before/index.js
@@ -4,7 +4,7 @@ const addEmptyLineBefore = require("../../utils/addEmptyLineBefore");
 const blockString = require("../../utils/blockString");
 const getPreviousNonSharedLineCommentNode = require("../../utils/getPreviousNonSharedLineCommentNode");
 const hasEmptyLine = require("../../utils/hasEmptyLine");
-const isAfterCommentLine = require("../../utils/isAfterCommentLine");
+const isAfterComment = require("../../utils/isAfterComment");
 const isCustomProperty = require("../../utils/isCustomProperty");
 const isFirstNested = require("../../utils/isFirstNested");
 const isSingleLineString = require("../../utils/isSingleLineString");
@@ -58,7 +58,7 @@ const rule = function(expectation, options, context) {
       // Optionally ignore the node if a comment precedes it
       if (
         optionsMatches(options, "ignore", "after-comment") &&
-        isAfterCommentLine(decl)
+        isAfterComment(decl)
       ) {
         return;
       }
@@ -78,7 +78,7 @@ const rule = function(expectation, options, context) {
         (optionsMatches(options, "except", "first-nested") &&
           isFirstNested(decl)) ||
         (optionsMatches(options, "except", "after-comment") &&
-          isAfterCommentLine(decl)) ||
+          isAfterComment(decl)) ||
         (optionsMatches(options, "except", "after-custom-property") &&
           isAfterCustomProperty(decl))
       ) {
