+++ lib/rules/declaration-empty-line-before/index.js
@@ -3,7 +3,7 @@
 const addEmptyLineBefore = require("../../utils/addEmptyLineBefore");
 const blockString = require("../../utils/blockString");
 const hasEmptyLine = require("../../utils/hasEmptyLine");
-const isAfterCommentLine = require("../../utils/isAfterCommentLine");
+const isAfterComment = require("../../utils/isAfterComment");
 const isAfterStandardPropertyDeclaration = require("../../utils/isAfterStandardPropertyDeclaration");
 const isCustomProperty = require("../../utils/isCustomProperty");
 const isFirstNested = require("../../utils/isFirstNested");
@@ -62,7 +62,7 @@ const rule = function(expectation, options, context) {
       // Optionally ignore the node if a comment precedes it
       if (
         optionsMatches(options, "ignore", "after-comment") &&
-        isAfterCommentLine(decl)
+        isAfterComment(decl)
       ) {
         return;
       }
@@ -90,7 +90,7 @@ const rule = function(expectation, options, context) {
         (optionsMatches(options, "except", "first-nested") &&
           isFirstNested(decl)) ||
         (optionsMatches(options, "except", "after-comment") &&
-          isAfterCommentLine(decl)) ||
+          isAfterComment(decl)) ||
         (optionsMatches(options, "except", "after-declaration") &&
           isAfterStandardPropertyDeclaration(decl))
       ) {
