+++ lib/utils/__tests__/isAfterComment.test.js
@@ -1,16 +1,16 @@
 /* flow */
 "use strict";
 
-const isAfterCommentLine = require("../isAfterCommentLine");
+const isAfterComment = require("../isAfterComment");
 const postcss = require("postcss");
 
-describe("isAfterCommentLine", () => {
+describe("isAfterComment", () => {
   it("returns true when after a single-line comment", () => {
     const root = postcss.parse(`
       /* comment */
       foo {}
     `);
-    expect(isAfterCommentLine(root.nodes[1])).toBe(true);
+    expect(isAfterComment(root.nodes[1])).toBe(true);
   });
 
   it("returns true when after a multi-line comment", () => {
@@ -19,7 +19,7 @@ describe("isAfterCommentLine", () => {
          and more comment */
       foo {}
     `);
-    expect(isAfterCommentLine(root.nodes[1])).toBe(true);
+    expect(isAfterComment(root.nodes[1])).toBe(true);
   });
 
   it("returns false when after a shared-line comment", () => {
@@ -27,7 +27,7 @@ describe("isAfterCommentLine", () => {
       bar {} /* comment */
       foo {}
     `);
-    expect(isAfterCommentLine(root.nodes[2])).toBe(false);
+    expect(isAfterComment(root.nodes[2])).toBe(false);
   });
 
   it("returns false when after a non-comment node", () => {
@@ -35,13 +35,13 @@ describe("isAfterCommentLine", () => {
       bar {}
       foo {}
     `);
-    expect(isAfterCommentLine(root.nodes[1])).toBe(false);
+    expect(isAfterComment(root.nodes[1])).toBe(false);
   });
 
   it("returns false when after no nodes", () => {
     const root = postcss.parse(`
       foo {}
     `);
-    expect(isAfterCommentLine(root.nodes[0])).toBe(false);
+    expect(isAfterComment(root.nodes[0])).toBe(false);
   });
 });
