+++ lib/utils/isAfterComment.js
